Rearrange DashboardVisualisation.ts
