+++ src/types/DashboardVisualisation.ts
@@ -1,29 +1,10 @@
 import { DataSourceID } from './DataSource';
 import { VisualisationType } from './Metadata';
-
-/**
- * Visualisation customisation
- */
-
-// This is the set of options for all visualisations
-export type BaseChartCustomisationOption = { description?: string; size: DashboardItemSize; paragraph?: string };
-
-// Add a customisation option for each visualisation possible
-export type LineChartCustomisationOption = BaseChartCustomisationOption;
-export type ThresholdChartCustomisationOption = BaseChartCustomisationOption & { thresholdValue?: number };
-
-// The union type of all available customisation option objects
-export type ChartCustomisationOption = LineChartCustomisationOption | ThresholdChartCustomisationOption;
+import { ChartCustomisationOption } from './VisualisationOption';
 
 /**
  * General
  */
-
-export enum DashboardItemSize {
-    SMALL = 2,
-    MEDIUM = 4,
-    LARGE = 6,
-}
 
 export type DashboardVisualisation = {
     dataSourceId: DataSourceID;
